+++ Packages/com.unity.inputsystem/InputSystem/Plugins/Steam/SteamIGAConverter.cs
@@ -26,7 +26,7 @@ using UnityEngine.InputSystem.Utilities;
 
 ////TODO: polling in background
 
-namespace UnityEngine.InputSystem.Plugins.Steam.Editor
+namespace UnityEngine.InputSystem.Steam.Editor
 {
     /// <summary>
     /// Converts input actions to and from Steam IGA file format.
@@ -83,7 +83,7 @@ namespace UnityEngine.InputSystem.Plugins.Steam.Editor
             builder.Append("using UnityEngine.InputSystem.Controls;\n");
             builder.Append("using UnityEngine.InputSystem.Layouts;\n");
             builder.Append("using UnityEngine.InputSystem.Utilities;\n");
-            builder.Append("using UnityEngine.InputSystem.Plugins.Steam;\n");
+            builder.Append("using UnityEngine.InputSystem.Steam;\n");
             builder.Append("#if UNITY_EDITOR\n");
             builder.Append("using UnityEditor;\n");
             builder.Append("#endif\n");