+++ Packages/com.unity.inputsystem/InputSystem/Plugins/UI/TouchModel.cs
@@ -5,7 +5,7 @@ using UnityEngine.EventSystems;
 using UnityEngine.InputSystem.Utilities;
 using UnityEngine.UI;
 
-namespace UnityEngine.InputSystem.Plugins.UI
+namespace UnityEngine.InputSystem.UI
 {
     internal struct TouchModel
     {