+++ Packages/com.unity.inputsystem/InputSystem/Plugins/UI/UIInputModule.cs
@@ -4,7 +4,7 @@ using UnityEngine.EventSystems;
 ////REVIEW: apparently EventSystem only supports a single "current" module so the approach here probably
 ////        won't fly and we'll have to roll all non-action modules into one big module
 
-namespace UnityEngine.InputSystem.Plugins.UI
+namespace UnityEngine.InputSystem.UI
 {
     /// <summary>
     /// Base class for <see cref="BaseInputModule">input modules</see> that send